New version with a split daemon and message queue.
cdr2.php is a drop-in replacement for cdr.php. It will stash the row data to be inserted in a message queue.
The daemon will receieve the queued message(s) and attempt to post it. If Google has a fault, the daemon will
retry several times before putting the message back in the queue and reinitializing the GDocs connection.